refactor(submissions): extract SubmissionItem to its own file and clean up imports
- Moves SubmissionItem to components/submissions/submission-item.tsx
- Removes unused and duplicate icon imports
- Fixes all linter errors
- Improves maintainability and potential for reuse
